[MOVE] #140 - FakeDataSource 패키지 정리
