+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/CartRepositoryImplTest.kt
@@ -5,10 +5,10 @@ import co.kr.woowahan_banchan.data.datasource.remote.detail.DetailDataSource
 import co.kr.woowahan_banchan.data.model.local.CartDto
 import co.kr.woowahan_banchan.data.model.remote.response.DetailDataResponse
 import co.kr.woowahan_banchan.data.model.remote.response.DetailResponse
-import co.kr.woowahan_banchan.data.repository.fakedatasource.FakeCartDataSource
-import co.kr.woowahan_banchan.data.repository.fakedatasource.FakeCartDataSourceWithError
-import co.kr.woowahan_banchan.data.repository.fakedatasource.FakeDetailDataSource
-import co.kr.woowahan_banchan.data.repository.fakedatasource.FakeDetailDataSourceWithError
+import co.kr.woowahan_banchan.data.repository.fakedatasource.local.cart.FakeCartDataSource
+import co.kr.woowahan_banchan.data.repository.fakedatasource.local.cart.FakeCartDataSourceWithError
+import co.kr.woowahan_banchan.data.repository.fakedatasource.remote.detail.FakeDetailDataSource
+import co.kr.woowahan_banchan.data.repository.fakedatasource.remote.detail.FakeDetailDataSourceWithError
 import co.kr.woowahan_banchan.domain.entity.cart.CartItem
 import co.kr.woowahan_banchan.domain.entity.error.ErrorEntity
 import co.kr.woowahan_banchan.domain.repository.CartRepository
