+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/fakedatasource/local/cart/FakeCartDataSource.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_banchan.data.repository.fakedatasource
+package co.kr.woowahan_banchan.data.repository.fakedatasource.local.cart
 
 import co.kr.woowahan_banchan.data.datasource.local.cart.CartDataSource
 import co.kr.woowahan_banchan.data.model.local.CartDto
