+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/fakedatasource/local/history/FakeHistoryDataSource.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_banchan.data.repository.fakedatasource
+package co.kr.woowahan_banchan.data.repository.fakedatasource.local.history
 
 import co.kr.woowahan_banchan.data.datasource.local.history.HistoryDataSource
 import co.kr.woowahan_banchan.data.model.local.HistoryDto
