+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/fakedatasource/local/order/FakeOrderDataSource.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_banchan.data.repository.fakedatasource
+package co.kr.woowahan_banchan.data.repository.fakedatasource.local.order
 
 import co.kr.woowahan_banchan.data.datasource.local.order.OrderDataSource
 import co.kr.woowahan_banchan.data.model.local.OrderDto
