+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/fakedatasource/remote/soupdish/FakeSoupDishDataSource.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_banchan.data.repository.fakedatasource
+package co.kr.woowahan_banchan.data.repository.fakedatasource.remote.soupdish
 
 import co.kr.woowahan_banchan.data.datasource.remote.soupdish.SoupDishDataSource
 import co.kr.woowahan_banchan.data.model.remote.response.DishResponse
